refac(zk): move `simple_evaluator` files to `expressions/evaluator`
