+++ tachyon/zk/plonk/circuit/expressions/BUILD.bazel
@@ -133,24 +133,6 @@ tachyon_cc_library(
     deps = [":expression"],
 )
 
-tachyon_cc_library(
-    name = "simple_evaluator",
-    hdrs = ["simple_evaluator.h"],
-    deps = [
-        ":advice_expression",
-        ":challenge_expression",
-        ":constant_expression",
-        ":evaluator",
-        ":fixed_expression",
-        ":instance_expression",
-        ":negated_expression",
-        ":product_expression",
-        ":scaled_expression",
-        ":selector_expression",
-        ":sum_expression",
-    ],
-)
-
 tachyon_cc_unittest(
     name = "expression_unittests",
     srcs = [
@@ -163,12 +145,10 @@ tachyon_cc_unittest(
         "product_expression_unittest.cc",
         "scaled_expression_unittest.cc",
         "selector_expression_unittest.cc",
-        "simple_evaluator_unittest.cc",
         "sum_expression_unittest.cc",
     ],
     deps = [
         ":expression_factory",
-        ":simple_evaluator",
         "//tachyon/math/elliptic_curves/bn/bn254:fr",
         "//tachyon/math/finite_fields/test:gf7",
         "//tachyon/math/polynomials/univariate:univariate_polynomial",
