+++ tachyon/zk/plonk/circuit/expressions/evaluator/simple_evaluator.h
@@ -4,8 +4,8 @@
 // can be found in the LICENSE-MIT.halo2 and the LICENCE-APACHE.halo2
 // file.
 
-#ifndef TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_SIMPLE_EVALUATOR_H_
-#define TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_SIMPLE_EVALUATOR_H_
+#ifndef TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_SIMPLE_EVALUATOR_H_
+#define TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_SIMPLE_EVALUATOR_H_
 
 #include <vector>
 
@@ -132,4 +132,4 @@ class SimpleEvaluator
 
 }  // namespace tachyon::zk
 
-#endif  // TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_SIMPLE_EVALUATOR_H_
+#endif  // TACHYON_ZK_PLONK_CIRCUIT_EXPRESSIONS_EVALUATOR_SIMPLE_EVALUATOR_H_
